deamon/requirements.txt
pycodestyle==2.12.1
flask==3.1.0
docker==7.1.0
python-dotenv==1.0.1
tinydb==4.8.2
